backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
python-dotenv==1.1.0
mysql-connector-python==9.3.0
gunicorn==23.0.0
Werkzeug==2.2.3